fix: pin base image to bookworm
Folks are not able to run the kitclient. The devcontainer build is breaking. It looks like there is a newer version of the base image that is incompatible with an option. Here is the error that makes me think that.

```
2025-11-14 01:20:15.983Z: #15 0.646 (!) The 'moby' option is not supported on Debian 'trixie' because 'moby-cli' and related system packages have been removed from that distribution.
2025-11-14 01:20:16.037Z: #15 0.647 (!) To continue, either set the feature option '"moby": false' or use a different base image (for example: 'debian:bookworm' or 'ubuntu-24.04').
```

So it looks like there are two choices: 1. pin to "bookworm" or 2. turn off the "moby" option.

This commit chooses the 1st option.

I don't know of a good way to test this without commiting to main. So here goes.